+++ src/Endpoints/Pipelines/IRetriever.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Endpoints.Pipelines
+{
+    public interface IRetriever<TIn, TOut>
+    {
+        Task<PipelineResponse<TOut>> Retrieve(TIn input);
+    }
+}
